requirements.txt
numpy==1.11.2
matplotlib==1.5.3
scikit_image==0.12.3
tensorflow==0.11.0rc1
Pillow==4.0.0
PyYAML==3.12
